requirements.txt
Pillow==10.0.1
Pillow==10.2.0
streamlit==1.31.1
